Don't run migration in production

+++ db/migrate/20230424153140_remove_old_user_data.rb
@@ -1,5 +1,7 @@
 class RemoveOldUserData < ActiveRecord::Migration[7.0]
   def change
+    return if Rails.env.production?
+
     remove_column :users, :bio
     remove_column :users, :roles
     remove_column :users, :insiders_status
